Order Event 관련 프로젝트 파일 구조 수정
