+++ order-service/src/main/java/com/shipflow/orderservice/domain/vo/CompanyInfo.java
@@ -1,5 +1,6 @@
 package com.shipflow.orderservice.domain.vo;
 
+
 import jakarta.persistence.Embeddable;
 import lombok.AccessLevel;
 import lombok.Getter;
