--- order-service/src/main/java/com/shipflow/orderservice/infrastructure/messaging/consume/ProductStockDecreasedEvent.java
@@ -1,16 +1,0 @@
-package com.shipflow.orderservice.infrastructure.messaging.consume;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.shipflow.common.messaging.event.SagaEvent;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class ProductStockDecreasedEvent extends SagaEvent {
-
-    private String orderId;
-    private String productId;
-    private Integer quantity;
-}
