+++ order-service/src/main/java/com/shipflow/orderservice/infrastructure/messaging/event/consume/ProductStockRestoredEvent.java
@@ -1,4 +1,4 @@
-package com.shipflow.orderservice.infrastructure.messaging.event.inbound;
+package com.shipflow.orderservice.infrastructure.messaging.event.consume;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.shipflow.common.messaging.event.SagaEvent;
@@ -10,7 +10,7 @@ import java.util.UUID;
 @Getter
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class ProductStockDecreasedFailedEvent extends SagaEvent {
+public class ProductStockRestoredEvent extends SagaEvent {
 
     private UUID orderId;
 }
