+++ order-service/src/main/java/com/shipflow/orderservice/infrastructure/messaging/event/consume/ShipmentCompletedEvent.java
@@ -1,16 +1,16 @@
-package com.shipflow.orderservice.infrastructure.messaging.consume;
+package com.shipflow.orderservice.infrastructure.messaging.event.consume;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.shipflow.common.messaging.event.SagaEvent;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.util.UUID;
+
 @Getter
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class ProductStockDecreasedFailedEvent extends SagaEvent {
+public class ShipmentCompletedEvent extends SagaEvent {
 
-    private String orderId;
-    private String productId;
-    private String reason;
+    private UUID orderId;
 }
