+++ order-service/src/main/java/com/shipflow/orderservice/infrastructure/messaging/event/consume/ShipmentCreatedEvent.java
@@ -1,15 +1,17 @@
-package com.shipflow.orderservice.infrastructure.messaging.consume;
+package com.shipflow.orderservice.infrastructure.messaging.event.consume;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.shipflow.common.messaging.event.SagaEvent;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.util.UUID;
+
 @Getter
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ShipmentCreatedEvent extends SagaEvent {
 
-    private String orderId;
-    private String shipmentId;
+    private UUID orderId;
+    private UUID shipmentId;
 }
