--- order-service/src/main/java/com/shipflow/orderservice/infrastructure/messaging/event/inbound/ProductStockRestoredEvent.java
@@ -1,16 +1,0 @@
-package com.shipflow.orderservice.infrastructure.messaging.event.inbound;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.shipflow.common.messaging.event.SagaEvent;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.UUID;
-
-@Getter
-@NoArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class ProductStockRestoredEvent extends SagaEvent {
-
-    private UUID orderId;
-}
