+++ order-service/src/main/java/com/shipflow/orderservice/infrastructure/messaging/event/publish/OrderCreationFailedEvent.java
@@ -1,6 +1,5 @@
-package com.shipflow.orderservice.infrastructure.messaging.event.outbound;
+package com.shipflow.orderservice.infrastructure.messaging.event.publish;
 
-import com.shipflow.common.messaging.event.EventType;
 import com.shipflow.common.messaging.event.SagaEvent;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,14 +8,16 @@ import java.util.UUID;
 
 @Getter
 @NoArgsConstructor
-public class OrderCreationStartedEvent extends SagaEvent {
+public class OrderCreationFailedEvent extends SagaEvent {
+
+    private static final String EVENT_TYPE = "order.creation.failed";
 
     private UUID orderId;
     private UUID productId;
     private int quantity;
 
-    public OrderCreationStartedEvent(UUID orderId, UUID productId, int quantity) {
-        super(EventType.ORDER_CREATION_STARTED);
+    public OrderCreationFailedEvent(UUID orderId, UUID productId, int quantity) {
+        super(EVENT_TYPE);
         this.orderId = orderId;
         this.productId = productId;
         this.quantity = quantity;
