+++ order-service/src/main/java/com/shipflow/orderservice/infrastructure/messaging/handler/ShipmentCreationFailedHandler.java
@@ -2,12 +2,12 @@ package com.shipflow.orderservice.infrastructure.messaging.handler;
 
 import com.shipflow.common.messaging.handler.AbstractSagaHandler;
 import com.shipflow.common.messaging.publisher.EventPublisher;
-import com.shipflow.config.message.RabbitMqConfig;
 import com.shipflow.orderservice.application.dto.OrderResult;
-import com.shipflow.orderservice.infrastructure.messaging.event.outbound.OrderCreationFailedEvent;
+import com.shipflow.orderservice.infrastructure.messaging.config.OrderRabbitConfig;
+import com.shipflow.orderservice.infrastructure.messaging.event.publish.OrderCreationFailedEvent;
 import com.shipflow.orderservice.application.service.OrderCommandService;
 import com.shipflow.orderservice.application.service.OrderQueryService;
-import com.shipflow.orderservice.infrastructure.messaging.event.inbound.ShipmentCreationFailedEvent;
+import com.shipflow.orderservice.infrastructure.messaging.event.consume.ShipmentCreationFailedEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.stereotype.Component;
@@ -20,7 +20,7 @@ public class ShipmentCreationFailedHandler extends AbstractSagaHandler<ShipmentC
     private final OrderQueryService orderQueryService;
     private final EventPublisher eventPublisher;
 
-    @RabbitListener(queues = RabbitMqConfig.QUEUE_ORDER_SHIPMENT_CREATION_FAILED)
+    @RabbitListener(queues = OrderRabbitConfig.QUEUE_ORDER_SHIPMENT_CREATION_FAILED)
     public void receive(ShipmentCreationFailedEvent event) {
         handle(event);
     }
