+++ order-service/src/main/java/com/shipflow/orderservice/infrastructure/messaging/handler/StockDecreasedHandler.java
@@ -1,9 +1,9 @@
 package com.shipflow.orderservice.infrastructure.messaging.handler;
 
 import com.shipflow.common.messaging.handler.AbstractSagaHandler;
-import com.shipflow.config.message.RabbitMqConfig;
 import com.shipflow.orderservice.application.service.OrderCommandService;
-import com.shipflow.orderservice.infrastructure.messaging.event.inbound.ProductStockDecreasedEvent;
+import com.shipflow.orderservice.infrastructure.messaging.config.OrderRabbitConfig;
+import com.shipflow.orderservice.infrastructure.messaging.event.consume.ProductStockDecreasedEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.stereotype.Component;
@@ -14,7 +14,7 @@ public class StockDecreasedHandler extends AbstractSagaHandler<ProductStockDecre
 
     private final OrderCommandService orderCommandService;
 
-    @RabbitListener(queues = RabbitMqConfig.QUEUE_ORDER_STOCK_DECREASED)
+    @RabbitListener(queues = OrderRabbitConfig.QUEUE_ORDER_STOCK_DECREASED)
     public void receive(ProductStockDecreasedEvent event) {
         handle(event);
     }
